ARM: OMAP: AM33XX: PM: Update sleep sequence for PG 2.0
PG 2.0 has fix for advisory 1.0.17 because of which we can now write to
REG_PHY_ENABLE_DYNAMIC_PWRDN to disable the input buffers.
This is done in u-boot as part of DDR configurations.

On enabling Dynamic Power-Down in PG2.0, the below workarounds can be
avoided in suspend sequence.
- switching to LVCMOS mode (setting MDDR_SEL in DDR_IO_CTRL)
- configuring pull downs on DATA0 and DATA1 macros

In this commit, switching to LVCMOS mode is skipped for PG2.0.

However, configuring pull-downs on DATA0 and DATA1 macros is not updated
here because after the change, it is observed that a higher VDDS_DDR
voltage was measured.

This commit has been tested on EVM 1.5, EVM, EVM-SK 1.2B

[gururaja.hebbar@ti.com]
	resolve merge conflicts

Signed-off-by: Satyanarayana Sandhya <sandhya.satyanarayana@ti.com>
Signed-off-by: Hebbar Gururaja <gururaja.hebbar@ti.com>