ARM: OMAP: AM33XX: PM: Add support for VTT Toggle
In order to keep the DDR in Self-Refresh, CKE pin needs to be Pulled
down.
The VTT regulator Pull-up is on the same CKE pin. The pull down enabled
on CKE is a weak pull down.  This weak pull-down is not enough to pull
the line below 0.75v.
Because of this, the DDR never properly goes to self refresh and hence
low power state is not achieved.

This issue is addressed as below on a few boards
1. Add an external pull down on CKE pin and
2. Provide a method to disable VTT regulator (via gpio toggling)

The 1st part of the solution is addressed in h/w.
This patch set handles the 2nd part of toggling the gpio in the
suspend/resume path (depending on whether this toggling is requested
from DT)

Note:
This VTT regulator should be disabled after enabling self-refresh on
suspend, and should be enabled before disabling self-refresh on resume.
This is to allow proper self refresh entry/exit commands to be
transmitted to the memory (VTT needs to be enabled whenever you perform
any addr/ctrl signaling)

Here it is assumed that VTT Toggle will be done using a pin on GPIO-0
Instance.

Signed-off-by: Hebbar Gururaja <gururaja.hebbar@ti.com>