ARM: dts: AM33XX: EVM-SK: Add support for VTT Toggle
Due to a Hardware related issue, Suspend-Resume causes a hangs on
EVM-SK board.

On EVM-SK, the CKE signal does not go low properly during DDR self
refresh. The pull down enabled on CKE is a weak pull down, which is not
enough to overcome the VTT pull up to 0.75V.  Thus, the DDR never
properly goes to self refresh. CKE needs to be pulled down to keep the
DDR in Self refresh.

EVM-SK (Rev 1.2) has an added external pull down on CKE and the ability
to disable VTT regulator via gpio to address the above issues.

AM335x PM code provides an option to toggle VTT which can be enabled
through DT.

Add necessary DT nodes to toggle VTT on EVM-SK board.
Also, add the required pin-mux setup for the gpio pins.

Signed-off-by: Hebbar Gururaja <gururaja.hebbar@ti.com>