ARM: OMAP: Wakeupgen: Add support for AM43XX
AM43XX has the same wakeupgen IP as OMAP4/5. The only
notable difference is the presence of 7 register banks
and lack of SAR area which has been used in OMAP4/5 for
saving and restoring the context around low power states.

In case of AM43XX the context is saved and restored by
the kernel.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>