i2c: omap: enhance pinctrl support
Update i2c-omap driver to set the state of the pins to:
- "default" on boot, resume and before performing an i2c transfer
- "idle" after initial default, after resume default, and after each
   i2c xfer
- "sleep" on suspend()

This should make it possible to optimize energy usage for the pins both
for the suspend/resume cycle, and for runtime cases in-between i2c
transfers.

By optionally putting the pins into default/idle/sleep state in the
suspend callback we can accomplish two things.
- minimize current leakage from pins and thus save power,
- prevent the IP from driving pins output in an uncontrolled manner,
   which may happen if the power domain drops the domain regulator.

Note:
- Changes based on spi-pl022.c.
- suspend()/resume() callbacks are added which simply puts the
  pins to sleep/default (& then to idle) state.

Signed-off-by: Hebbar, Gururaja <gururaja.hebbar@ti.com>